chore(dashboard): trim layoutConfig to live geometry; delete kpiQueries + config orphans
- layoutConfig: drop WIDGETS/DEFAULT_LAYOUT/the dash-case helper functions and the
  supervisorMetrics/kpiDisplay/kpiSparkline imports; keep only the live geometry
  (GRID_COLS, KPI_ROW_HEIGHT, GRID_MARGIN_Y, FULL_WIDTH_TABLE_GRID, UNIFORM/MAP
  constraints, DEFAULT_CHART_LAYOUT, findFirstOpenPosition).
- AddKpiDropdown: drop the legacy INVENTORY_* fallback (catalogItems is always
  supplied now); list comes purely from the catalog.
- Extract formatDimensionLabel (the only live export of the ~3000-line kpiQueries)
  into config/labelFormat.js; repoint mapHoverPresentation/DashboardHeader/mapGeoUtils;
  DELETE config/kpiQueries.js.
- That orphaned + DELETED: complaintAgeBucketConfig, complaintChannelConfig,
  complaintTypeDepartmentConfig, kpiSparkline.

Dev build compiles clean; 11 tiles render.
